Add ability to add juttle engine to an express app.
Split juttle-engine.js into two pieces. All the express route handlers
are now in express-routes.js. The main function takes an express app
and options and attaches all the routes to the app. We use an express
router to hold the middleware (bodyParser), error
handling (default_error_handler), and all of the http-based routes.

However, the package we use for websocket handling (express-ws)
doesn't support adding endpoints to routers, only apps. For that
reason, the main function adds to an app instead of returning a
standalone router.

juttle-engine.js now just creates the express app, uses express-routes
to add the routes to the app, and starts the server.

Also switch to ES6 classes instead of extendable-base.