Improve test coverage.
This change caused the coverage test percentages to dip slightly below
the thresholds. I noticed that there wasn't any test for the
rendezvous functionality, so add a test.

We'll probably move this to the standalone workbench app once it
exists, but until we do, we should keep the test to keep coverage.