juttle-service: refactor the service module to be embeddable
Change around the juttle-service module so it exports functions
for configure, add_routes, and run. This way an embedding server
like juttle-engine can just use the configure / add_routes bits
and not need to interact with the class.

Fix package.json so it exports the service class and not the
bin/juttle-service script so that `require('juttle-service')`
works properly.