Don't confuse compile errors with other errors.
In some cases, the act of compiling and running a juttle program might
result in an error being thrown that is not one of the JuttleError
types. In that case, juttle-subprocess was returning a compile_error
message with the error object as an argument, which was misleading.

Fix this by:

 - Changing the name of the message for juttle errors from
   'compile_error' to 'juttle_error'.

 - Adding a new message type from the juttle subprocess
   'internal_error' that is returned when a non-juttle error is
   caught.

In juttle-job, juttle_errors are rejected as juttleErrors (resulting
in a 400), while internal_errors are rejected as
unknownError(resulting in a 500).