Make juttle-service-client a module.
Move most of the implementation of juttle-service-client a module so
it can be easily re-used in juttle-engine-client, which has most of
the same capabilities.

This module exports the following functions:

 - init: override the name of the option that specifies the
   daemon (juttle-service or juttle-engine). Can also provide
   additional commands and/or options (this will be used to add the
   rendezvous options when used by juttle-engine).
 - usage: usage for the program.
 - handle_command: the switch block that does all the work based on command.

juttle-service-client is now limited to option capturing and calling handle_command.

This also removes the rendezvous functionality. We'll add it back in
juttle-engine, which has a app to which we want to push programs/watch
programs.