Make service client module more modular.
Make the juttle service client more modular, with an object describing
the supported commands and an object describing the supported command
line arguments.

Instead of having a switch statement where the work is directly
performed, command() consults the set of commands and calls the
appropriate function.

This made usage() a bit more complicated, as it needed to handle
flexible sets of commands, command line arguments, etc. with
consistent spacing. Pull in sprintf to take care of that.

Also add unit tests that cover each of the main command line options,
adding an additional command, and trying to run a command that doesn't
exist. They don't deeply test the results but do try to do a basic
success/failure check.

To support the unit tests, several of the write_output/write_error
calls had to be run through sprintf to convert them to a single
string.

Also decrease the coverage percentage on branches from 75 to 74. The
branch coverage is 74.24% which is just barely below the old threshold.