Add unit tests for logSetup().
We're basically at the minimum branch coverage percentage, and
log-setup.js was one of the least covered files, so add unit
tests. They cover the following:

 - Providing log config directly to log4js.
 - Specifying a log-level independent from log4js.
 - Specifying a log file independent from log4js.
 - Using DEBUG= in the environment to specify debug log components.
 - Throwing errors when a logfile cannot be created or written to.
 - Uses the default output file when daemonizing.

While doing this, I found a race condition in the way I was attempting
to create log files if they didn't exist. fs-extra's ensureFileSync is
simpler, so use that instead.