Add a /config-info endpoint, revamp config management.
Add a /config-info endpoint that returns the configuration of the
juttle-service. To manage this correctly, this led to a bunch of other
changes that rearranged how config was stored and how it can be
overridden.

The config is now created in
juttle-service.js:configure. Configuration is applied in the following
precedence order:

 - the const DEFAULT_CONFIG
 - .juttle-config.json (obtained from read_config()) under the "juttle-service" property
 - configuration provided directly, either directly for unit tests or
   via command line args in the juttle-service binary.

configure now also fully resolves the root path after all the above
merges.

We still have to pass the config file path around so it can be used
when invoking the juttle subprocess. Pass this explicitly instead of
in a general options object. Generally, be more consistent about using
the variable config to hold config and opts to hold command line
arguments.

Add the new endpoint to the the jobs api doc.

Add unit tests that test both the precedence of the overrides as well
as fetching the /config endpoint to fetch the config.

Update a few unit tests that were passing direct config as items below
a config: juttle-service: property to just provide them directly.

I also was running across a few failures with the
juttle-service-client tests timing out. The default wait time of 1
second was a bit too aggressive, as I was seeing tests taking ~600 on
average. Increased the timeout up to 5 seconds and made it a constant
shared across all the tests.

Update the README to reflect this precedence. Also fix some out of
date items (wrong log path, missing delayed_endpoint_close).